% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fazhthemes.R
\docType{package}
\name{fazhthemes}
\alias{fazhthemes}
\title{fazhthemes: A package with {ggplot2} themes.}
\description{
The {fazhthemes} package provides the author's (Fernando Antonio Zepeda Herrera) personal
themes for plotting in R.
}
\section{Lucified themes}{

The lucify_theme_* functions are modified versions of {ggplot2}'s core themes.
}

