% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucify_themes.R
\name{lucify_basics}
\alias{lucify_basics}
\title{Lucified ggplot2 themes}
\usage{
lucify_basics(
  text_family = "Century Gothic",
  text_size = 17,
  text_color = "#3C3C3C"
)
}
\arguments{
\item{text_family}{(Default = "Century Gothic") Text font family to use.}

\item{text_size}{(Default = 17) Text size to use.}

\item{text_color}{(Default = "#3C3C3C") Text color to use.}
}
\description{
These functions "lucify" a core ggplot2 theme.
The basics of lucified themes are Century Gothic fonts, bigger sizes of text elements and margins,
as well as slightly different colors and centered title and subtitles.
}
\examples{
\dontrun{
library(tidyverse)
UKDriverDeaths_tibble <- UKDriverDeaths \%>\%
  matrix(nrow=16, ncol=12, byrow = TRUE,
         dimnames = list(1969:1984,month.abb)) \%>\%
  as.data.frame() \%>\%
  rownames_to_column("Month")

ggplot(data = ULDriverDeaths_tibble, aes(x=Month,y=Jan)) +
  geom_col(fill = "steelblue4", width = 0.1) +
  geom_point(color = "steelblue4", size = rel(10)) +
  theme_minimal() +
  lucify_basics()
}

}
